+++ src/d_netcmd.h
@@ -76,7 +76,7 @@ extern consvar_t cv_teamplay;
 extern consvar_t cv_teamdamage;
 extern consvar_t cv_fraglimit;
 extern consvar_t cv_timelimit;
-extern ULONG timelimitintics;
+extern uint32_t timelimitintics;
 extern consvar_t cv_allowturbo;
 extern consvar_t cv_allowexitlevel;
